requirements.txt
# Korean Food Agent Dependencies

# ===========================
# LangGraph & LangChain (필수)
# ===========================
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# ===========================
# FastAPI & Web (필수)
# ===========================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ===========================
# Database (필수)
# ===========================
supabase>=2.0.0

# ===========================
# Web Scraping (필수)
# ===========================
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
nest-asyncio>=1.5.0

# ===========================
# Image Processing (필수)
# ===========================
Pillow>=10.0.0

# ===========================
# Utilities (필수)
# ===========================
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
requests>=2.31.0
aiofiles>=23.0.0
